embedding/requirements.txt
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
clean-text==0.6.0
contourpy==1.2.0
cycler==0.12.1
emoji==1.7.0
fonttools==4.49.0
ftfy==6.1.3
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
kmedoids==0.5.0
llvmlite==0.42.0
matplotlib==3.8.3
mealpy==3.0.1
numba==0.59.0
numpy==1.26.4
opfunu==1.0.1
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyarrow==15.0.0
pynndescent==0.5.11
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
reportlab==4.1.0
requests==2.31.0
scikit-learn==1.4.1.post1
scikit-learn-extra==0.3.0
scipy==1.12.0
six==1.16.0
threadpoolctl==3.3.0
tqdm==4.66.2
tzdata==2024.1
umap-learn==0.5.5
urllib3==2.2.1
wcwidth==0.2.13
